requirements.txt
langchain
langchain_openai
langchain_pinecone
pinecone_client
streamlit
unstructured
unstructured[pdf]
openai
chromadb
tiktoken
langchain-community
psutil
python-magic
python-pptx



